refactor: 카드 브랜드 및 카드사 타입과 상수를 새로운 파일로 분리
